feat: move atomic components to ui folder
