+++ src/components/credentialList.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { useIntl } from "react-intl";
 import { CredentialCard } from "@components/credentialCard";
-import { Loader } from "@components/loader";
+import { Loader } from "@components/ui";
 import { IMessage } from "@pages/background/types";
 
 export function CredentialList(): JSX.Element {
