+++ src/components/ui/index.ts
@@ -3,3 +3,6 @@ export * from "./drawer";
 export * from "./typography";
 export * from "./card";
 export * from "./dropdown";
+export * from "./radio";
+export * from "./switch";
+export * from "./loader";
