+++ src/components/ui/loader/index.ts
@@ -1,0 +1,1 @@
+export { Loader } from "./loader";
