+++ src/components/ui/radio/index.ts
@@ -1,0 +1,1 @@
+export { Radio } from "./radio";
