+++ src/pages/popup/Popup.tsx
@@ -6,7 +6,7 @@ import { LocaleProvider } from "@src/_locales";
 import { default as defaultVendor } from "@src/config/vendor.json";
 import { IMessage } from "@pages/background/types";
 import { Signin } from "@src/screens/signin";
-import { Loader } from "@components/loader";
+import { Loader } from "@components/ui";
 import { Main } from "@components/main";
 
 interface IConnect {
